Moved the support landing ab test files into a subfolder inside bundles landing page.
