+++ assets/stylesheets/main.scss
@@ -47,7 +47,7 @@
 // ----- Pages ----- //
 
 @import '../pages/bundles-landing/bundlesLanding';
-@import '../pages/bundles-landing/supportLanding';
+@import '../pages/bundles-landing/support-landing-ab-test/supportLanding';
 @import '../pages/contributions-landing/contributionsLanding';
 @import '../pages/regular-contributions/regularContributions';
 @import '../pages/oneoff-contributions/oneoffContributions';
